Reduce L to get better clustering in single trap.
If L=5000, when a cluster forms within a trap, this makes a gradient
that makes another cluster form outside the trap. This is messy and
I don't want this. Reducing L=5000 to L=2500 means there is not space
for this to happen, and still leaves space for the multiple trap
experiment.