Stop finding `c` when doing trap run for `chi`=0.
When doing `run_trap_nochi`, where we want to find the fraction of
particles in a trap when there is no chemotaxis, just due to wall
alignment, there's no need to find the chemical field. So set
`c_source` to zero so that the model knows not to iterate it.