Fix traps so that physical lengths are accurate.
Before, if `w_i` or `s_i` were odd, the extra `1` did not make
any difference to the trap made. But the physical lengths `w` and `s`
were not amended to reflect this.

I've also added docstrings to make clear the valid values that will
actually result in the desired lengths.